restructure part 1
